Match fcl misc narrow context id

+++ src/Event/Fcl/fcl_misc.c
@@ -3489,12 +3489,11 @@ void FUN_003cda60(u64 param_1)
 // FUN_003CDBA0 NONMATCHING
 
 
-u32 FUN_003cdba0(u64 param_1,u32 param_2)
+u32 FUN_003cdba0(u64 param_1,u16 param_2)
 {
   u32 result;
   s32 context;
   u32 flags;
-
   context = fclMiscContextCall();
   if (context == 0) {
     fclMiscAssertCall((u32)DAT_006a3e18,0x9e2);
